refactor: move localstorage code to ui package
